# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Structured Cabling Labor Estimator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /estimates:
    get:
      operationId: listEstimates
      tags: [estimates]
      summary: List all estimates
      responses:
        "200":
          description: List of estimates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EstimateSummary"
    post:
      operationId: createEstimate
      tags: [estimates]
      summary: Create a new estimate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEstimateBody"
      responses:
        "201":
          description: Created estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"
  /estimates/{id}:
    get:
      operationId: getEstimate
      tags: [estimates]
      summary: Get an estimate with all runs and computed totals
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Estimate with runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateDetail"
        "404":
          description: Not found
    put:
      operationId: updateEstimate
      tags: [estimates]
      summary: Update an estimate
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEstimateBody"
      responses:
        "200":
          description: Updated estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"
    delete:
      operationId: deleteEstimate
      tags: [estimates]
      summary: Delete an estimate
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /estimates/{id}/hardware-items:
    post:
      operationId: createCablingHardwareItem
      tags: [hardwareItems]
      summary: Add a hardware/material line item to a cabling estimate
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHardwareItemBody"
      responses:
        "201":
          description: Created hardware item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HardwareItem"
  /hardware-items/{id}:
    put:
      operationId: updateCablingHardwareItem
      tags: [hardwareItems]
      summary: Update a cabling hardware/material line item
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHardwareItemBody"
      responses:
        "200":
          description: Updated hardware item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HardwareItem"
    delete:
      operationId: deleteCablingHardwareItem
      tags: [hardwareItems]
      summary: Delete a cabling hardware/material line item
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    InstallType:
      type: string
      enum: [new_install, retrofit, deinstall]
    CeilingType:
      type: string
      enum: [open, drywall, hard_lid]
    PathwayComplexity:
      type: string
      enum: [low, medium, high]
    BuildingType:
      type: string
      enum: [office, warehouse, retail, healthcare]
...
    WorkEnvironment:
      type: string
      enum: [occupied, unoccupied]
    SkillLevel:
      type: string
      enum: [apprentice, journeyman, lead]
    EstimateSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        installType:
          $ref: "#/components/schemas/InstallType"
        buildingType:
          $ref: "#/components/schemas/BuildingType"
        environment:
          $ref: "#/components/schemas/WorkEnvironment"
        skillLevel:
          $ref: "#/components/schemas/SkillLevel"
        hourlyRate:
          type: number
        runCount:
          type: integer
        totalDrops:
          type: integer
        totalHoursAvg:
          type: number
        totalCostAvg:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        [
          id,
          name,
          installType,
          buildingType,
          environment,
          skillLevel,
          hourlyRate,
          runCount,
          totalDrops,
          totalHoursAvg,
          totalCostAvg,
          createdAt,
          updatedAt,
        ]
    Estimate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        installType:
          $ref: "#/components/schemas/InstallType"
        buildingType:
          $ref: "#/components/schemas/BuildingType"
        environment:
          $ref: "#/components/schemas/WorkEnvironment"
        skillLevel:
          $ref: "#/components/schemas/SkillLevel"
        hourlyRate:
          type: number
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        [
          id,
          name,
          installType,
          buildingType,
          environment,
          skillLevel,
          hourlyRate,
          createdAt,
          updatedAt,
        ]
...
    RunCalculation:
      type: object
      description: Computed labor breakdown for a single cable run group
      properties:
        runId:
          type: integer
        label:
          type: string
        cableType:
          type: string
        description:
          type: string
          nullable: true
        numCables:
          type: integer
          description: Cables pulled simultaneously (B) — drives the bulk factor
        fiberStrands:
          type: integer
          description: Strands per fiber cable; multiplies termination labor (1 for non-fiber)
        lengthFt:
          type: number
        ceilingType:
          $ref: "#/components/schemas/CeilingType"
        pathwayComplexity:
          $ref: "#/components/schemas/PathwayComplexity"
        bulkFactor:
          type: number
          description: "Computed efficiency factor: 0.4 + (0.6 / numCables)"
        pullMinutesPer10Ft:
          type: number
          description: Base pull rate in minutes per 10 ft for this cable type
        terminationMinutesPerEnd:
          type: number
          description: Base termination time in minutes per end for this cable type
        conditionMultiplier:
          type: number
          description: Combined site/condition multiplier (install × ceiling × pathway × building × environment × skill)
        pullHoursPerCable:
          type: number
          description: Pull hours for one cable within a single bulk pass (after condition mult + bulk factor)
        terminationHoursPerCable:
          type: number
          description: Termination hours per cable (both ends, after condition mult — NOT bulk-discounted)
        adjustedHoursPerCable:
          type: number
          description: Average total hours per cable = (totalPullHours + totalTermHours) / numCables
        runHoursLow:
          type: number
        runHoursAvg:
          type: number
        runHoursHigh:
          type: number
        runCostLow:
          type: number
        runCostAvg:
          type: number
        runCostHigh:
          type: number
      required:
        [
          label,
          cableType,
          numCables,
          lengthFt,
          ceilingType,
          pathwayComplexity,
          bulkFactor,
          pullMinutesPer10Ft,
          terminationMinutesPerEnd,
          conditionMultiplier,
          pullHoursPerCable,
          terminationHoursPerCable,
          adjustedHoursPerCable,
          runHoursLow,
          runHoursAvg,
          runHoursHigh,
          runCostLow,
          runCostAvg,
          runCostHigh,
        ]
    TaskBreakdownItem:
      type: object
      properties:
        task:
          type: string
        percent:
          type: number
        hoursAvg:
          type: number
        costAvg:
          type: number
      required: [task, percent, hoursAvg, costAvg]
    EstimateTotals:
      type: object
      properties:
        totalCables:
          type: integer
        totalRuns:
          type: integer
        totalHoursLow:
          type: number
        totalHoursAvg:
          type: number
        totalHoursHigh:
          type: number
        totalCostLow:
          type: number
        totalCostAvg:
          type: number
        totalCostHigh:
          type: number
        bulkSavingsHours:
          type: number
          description: Hours saved by pulling cables in a single bulk pass vs. pulling each one solo
        taskBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/TaskBreakdownItem"
      required:
        [
          totalCables,
          totalRuns,
          totalHoursLow,
          totalHoursAvg,
          totalHoursHigh,
          totalCostLow,
          totalCostAvg,
          totalCostHigh,
          bulkSavingsHours,
          taskBreakdown,
        ]
    EstimateDetail:
      type: object
      properties:
        estimate:
          $ref: "#/components/schemas/Estimate"
        runs:
          type: array
          items:
            $ref: "#/components/schemas/RunCalculation"
        totals:
          $ref: "#/components/schemas/EstimateTotals"
        hardwareItems:
          type: array
          items:
            $ref: "#/components/schemas/HardwareItem"
        materials:
          $ref: "#/components/schemas/MaterialBreakdown"
        projectTotal:
          type: number
          description: totals.totalCostAvg + materials.total
      required: [estimate, runs, totals, hardwareItems, materials, projectTotal]
    CreateEstimateBody:
      type: object
      properties:
        name:
          type: string
        installType:
          $ref: "#/components/schemas/InstallType"
        buildingType:
          $ref: "#/components/schemas/BuildingType"
        environment:
          $ref: "#/components/schemas/WorkEnvironment"
        skillLevel:
          $ref: "#/components/schemas/SkillLevel"
        hourlyRate:
          type: number
        notes:
          type: string
          nullable: true
      required:
        [name, installType, buildingType, environment, skillLevel, hourlyRate]
    UpdateEstimateBody:
      type: object
      properties:
        name:
          type: string
        installType:
          $ref: "#/components/schemas/InstallType"
        buildingType:
          $ref: "#/components/schemas/BuildingType"
        environment:
          $ref: "#/components/schemas/WorkEnvironment"
        skillLevel:
          $ref: "#/components/schemas/SkillLevel"
        hourlyRate:
          type: number
        notes:
          type: string
          nullable: true
      required:
        [name, installType, buildingType, environment, skillLevel, hourlyRate]
...
    HardwareItem:
      type: object
      description: A hardware/material line item attached to an estimate.
      properties:
        id: { type: integer }
        estimateId: { type: integer }
        catalogKey:
          type: string
          nullable: true
          description: If set, references a hardwareCatalog entry by its `value`. May be null for one-off items.
        name: { type: string }
        quantity: { type: number, minimum: 0 }
        unitCost: { type: number, minimum: 0 }
        unit:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        sortOrder: { type: integer }
        lineTotal:
          type: number
          description: quantity × unitCost
        createdAt:
          type: string
          format: date-time
      required:
        [
          id,
          estimateId,
          name,
          quantity,
          unitCost,
          sortOrder,
          lineTotal,
          createdAt,
        ]
    CreateHardwareItemBody:
      type: object
      properties:
        catalogKey:
          type: string
          nullable: true
          maxLength: 200
        name: { type: string, minLength: 1, maxLength: 200 }
        quantity: { type: number, minimum: 0, maximum: 1000000 }
        unitCost: { type: number, minimum: 0, maximum: 1000000 }
        unit:
          type: string
          nullable: true
          maxLength: 50
        notes:
          type: string
          nullable: true
          maxLength: 1000
      required: [name, quantity, unitCost]
    UpdateHardwareItemBody:
      type: object
      properties:
        catalogKey:
          type: string
          nullable: true
          maxLength: 200
        name: { type: string, minLength: 1, maxLength: 200 }
        quantity: { type: number, minimum: 0, maximum: 1000000 }
        unitCost: { type: number, minimum: 0, maximum: 1000000 }
        unit:
          type: string
          nullable: true
          maxLength: 50
        notes:
          type: string
          nullable: true
          maxLength: 1000
      required: [name, quantity, unitCost]
    MaterialCableLine:
      type: object
      properties:
        cableType: { type: string }
        totalLengthFt: { type: number }
        totalCables: { type: integer }
        costPerFt: { type: number }
        subtotal: { type: number }
      required: [cableType, totalLengthFt, totalCables, costPerFt, subtotal]
    MaterialTerminationLine:
      type: object
      properties:
        cableType: { type: string }
        totalEnds: { type: integer }
        costPerEnd: { type: number }
        subtotal: { type: number }
      required: [cableType, totalEnds, costPerEnd, subtotal]
    MaterialBreakdown:
      type: object
      description: |
        Material cost breakdown for an estimate. For Cabling estimates, cable + termination
        subtotals are auto-computed from runs. For Pathway estimates, pathwaySubtotal is
        the sum of segment material (already includes pathway, fasteners, bends, penetrations).
        Hardware items are user-added line items. Waste is applied to cableSubtotal only.
        Markup is applied to the full subtotal.
      properties:
        cableSubtotal: { type: number }
        cableWasteAmount: { type: number }
        terminationSubtotal: { type: number }
        pathwaySubtotal: { type: number }
        hardwareSubtotal: { type: number }
        subtotal: { type: number, description: "Sum of (cable + waste) + termination + pathway + hardware" }
        wastePercent: { type: number }
        markupPercent: { type: number }
        markupAmount: { type: number }
        total: { type: number, description: "subtotal + markupAmount" }
        cableLines:
          type: array
          items: { $ref: "#/components/schemas/MaterialCableLine" }
        terminationLines:
          type: array
          items: { $ref: "#/components/schemas/MaterialTerminationLine" }
      required:
        [
          cableSubtotal,
          cableWasteAmount,
          terminationSubtotal,
          pathwaySubtotal,
          hardwareSubtotal,
          subtotal,
          wastePercent,
          markupPercent,
          markupAmount,
          total,
          cableLines,
          terminationLines,
        ]
...