# lib/api-zod/src/generated/types/createHardwareItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Structured Cabling Labor Estimator API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateHardwareItemBody {
  /** @maxLength 200 */
  catalogKey?: string | null;
  /**
   * @minLength 1
   * @maxLength 200
   */
  name: string;
  /**
   * @minimum 0
   * @maximum 1000000
   */
  quantity: number;
  /**
   * @minimum 0
   * @maximum 1000000
   */
  unitCost: number;
  /** @maxLength 50 */
  unit?: string | null;
  /** @maxLength 1000 */
  notes?: string | null;
}