# lib/api-zod/src/generated/types/estimateDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Structured Cabling Labor Estimator API
 * OpenAPI spec version: 0.1.0
 */
import type { Estimate } from "./estimate";
import type { EstimateTotals } from "./estimateTotals";
import type { HardwareItem } from "./hardwareItem";
import type { MaterialBreakdown } from "./materialBreakdown";
import type { RunCalculation } from "./runCalculation";

export interface EstimateDetail {
  estimate: Estimate;
  runs: RunCalculation[];
  totals: EstimateTotals;
  hardwareItems: HardwareItem[];
  materials: MaterialBreakdown;
  /** totals.totalCostAvg + materials.total */
  projectTotal: number;
}