# lib/api-zod/src/generated/types/hardwareItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Structured Cabling Labor Estimator API
 * OpenAPI spec version: 0.1.0
 */

/**
 * A hardware/material line item attached to an estimate.
 */
export interface HardwareItem {
  id: number;
  estimateId: number;
  /** If set, references a hardwareCatalog entry by its `value`. May be null for one-off items. */
  catalogKey?: string | null;
  name: string;
  /** @minimum 0 */
  quantity: number;
  /** @minimum 0 */
  unitCost: number;
  unit?: string | null;
  notes?: string | null;
  sortOrder: number;
  /** quantity × unitCost */
  lineTotal: number;
  createdAt: Date;
}