# lib/api-zod/src/generated/types/materialCableLine.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Structured Cabling Labor Estimator API
 * OpenAPI spec version: 0.1.0
 */

export interface MaterialCableLine {
  cableType: string;
  totalLengthFt: number;
  totalCables: number;
  costPerFt: number;
  subtotal: number;
}